refactor: Minor changes on props
